+++ packages/resources/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @serverless-stack/resources
 
+## 1.17.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @serverless-stack/core@1.17.1
+
 ## 1.17.0
 
 ### Minor Changes